pyproject.toml
[project]
name = "nilotica"
dynamic = ["version"]
description = "一个基于vnpy的量化系统(Based vnpy quant trading system.)"
license = { file = "LICENSE" }
authors = [{ name = "Donny", email = "donnymoving@gmail.com" }]
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Office/Business :: Financial :: Investment",
    "Natural Language :: Chinese (Simplified)",
    "Typing :: Typed",
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "deap>=1.4.2",
    "msgpack==1.1.0",
    "numpy>=2.2.3",
    "peewee==3.17.9",
    "plotly>=6.0.0",
    "pybind11>=2.9.0",
    "pybind11-stubgen>=2.5.3",
    "pyqtgraph>=0.13.7",
    "pyside6>=6.8.2.1",
    "pyzmq>=26.3.0",
    "qdarkstyle>=3.2.3",
    "seaborn==0.13.2",
    "tqdm>=4.67.1",
    "loguru>=0.7.3",
    "nbformat>=5.10.4",
    "tushare==1.4.21",
    "tzlocal>=5.3.1",
    "pandas>=2.2.3",
    "hatch>=1.14.1",
    "setuptools>=78.1.0",
    "tzdata>=2025.2",
    "pathlib>=1.0.1",
    "pip>=25.0.1",
    "hatchling>=1.27.0",
    "babel>=2.17.0",
    "polars>=1.26.0",
    "websocket-client>=1.8.0",
    "importlib-metadata>=8.6.1",
    "pyyaml>=6.0.2",
    "pydantic>=2.11.4",
]

keywords = ["quant", "quantitative", "trading", "nilotica", "investment"]

[project.urls]
"Homepage" = "https://github.com/Lumosylva/Nilotica"
"Documentation" = "https://github.com/Lumosylva/Nilotica/README.md"
"Repository" = "https://github.com/Lumosylva/Nilotica.git"
"Issues" = "https://github.com/Lumosylva/Nilotica/issues"
"Changes" = "https://github.com/Lumosylva/Nilotica/CHANGELOG.md"
"Source" = "https://github.com/Lumosylva/Nilotica"

[build-system]
requires = [
    "hatchling>=1.27.0",
    "babel>=2.17.0",
    "pybind11>=2.13.6",
    "pybind11-stubgen>=2.5.3",
    "setuptools",
    "numpy",
    "tzdata>=2025.2",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "__init__.py"
pattern = "__version__ = ['\"](?P<version>[^'\"]+)['\"]"

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"
#path = "hatch_build_tts.py" # for tts

[tool.hatch.build.targets.sdist]
include = [
    "config*",
    "locales*",
    "ta-lib*",
    "utils*",
    "vnpy*",
    "vnpy_ctp*",
    "vnpy_tts*",
    "vnpy_websocket*",
    "zmq_services*",
    "__init__.py",
    ".python-version",
    "babel.cfg",
    "build.bat",
    "build.sh",
    "CHANGELOG.md",
    "CODE_OF_CONDUCT.md",
    "extract_logger.py",
    "hatch_build.py",
    "hatch_build_tts.py",
    "install_ta_lib.bat",
    "install_ta_lib.sh",
    "LICENSE",
    "locales.bat",
    "locales.sh",
    "main.py",
    "pyproject.toml",
    "README.md",
    "README_EN.md",
    "run.bat",
    "run.sh",
    "vnpy_ctp/api/**/*.h",
    "vnpy_ctp/api/**/*.cpp",
    "vnpy_ctp/api/**/*.lib",
    "vnpy_ctp/api/**/*.a",
    "vnpy_tts/api/**/*.h",
    "vnpy_tts/api/**/*.cpp",
    "vnpy_tts/api/**/*.lib",
    "vnpy_tts/api/**/*.a",
]

[tool.hatch.build.targets.wheel]
packages = [
    "config",
    "logger",
    "structlog",
    "utils",
    "vnpy",
    "vnpy_ctp",
    "vnpy_tts",
    "vnpy_websocket",
    "zmq_services",
    "__init__.py",
    ".python-version",
    "babel.cfg",
    "build.bat",
    "build.sh",
    "CHANGELOG.md",
    "CODE_OF_CONDUCT.md",
    "extract_logger.py",
    "hatch_build.py",
    "hatch_build_tts.py",
    "install_ta_lib.bat",
    "install_ta_lib.sh",
    "LICENSE",
    "locales.bat",
    "locales.sh",
    "main.py",
    "pyproject.toml",
    "README.md",
    "README_EN.md",
    "run.bat",
    "run.sh",
    "vnpy_ctp/api/**/*.h",
    "vnpy_ctp/api/**/*.cpp",
    "vnpy_ctp/api/**/*.lib",
    "vnpy_ctp/api/**/*.a",
    "vnpy_tts/api/**/*.h",
    "vnpy_tts/api/**/*.cpp",
    "vnpy_tts/api/**/*.lib",
    "vnpy_tts/api/**/*.a",
]

[project.optional-dependencies]
dev = [
    "hatchling>=1.27.0",
    "babel>=2.17.0",
    "scipy>=1.15.2",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "isort>=5.13.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
]

[tool.hatch.envs.default]
dependencies = [
    "black>=23.12.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "nilotica @ ."
]

scripts.test = "pytest {args}"
scripts.lint = [
    "black . --include \"*.py\"",
    "isort . --profile black",
    "mypy -p nilotica --ignore-missing-imports",
    "ruff check .",
]
scripts.format = [
    "black . --include \"*.py\"",
    "isort . --profile black",
]
scripts.clean = "(rd /s /q .pytest_cache & rd /s /q .mypy_cache & rd /s /q .ruff_cache & rd /s /q dist & rd /s /q build & for /d %G in (*.egg-info) do @rd /s /q \"%G\")"
scripts.build = "hatch build"

[tool.hatch.envs.test]
dependencies = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "nilotica @ ."
]
scripts.run = "pytest {args:tests}"

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.12.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "isort>=5.13.0",
]
scripts.style = "black --check --diff ."
scripts.typing = "mypy ."
scripts.lint = [
    "style",
    "typing",
    "ruff check ."
]
scripts.format = [
    "black .",
    "isort .",
]

[tool.ruff]
target-version = "py312"
output-format = "full"
line-length = 120

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]
ignore = ["E501"]  # 忽略行长度限制

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=nilotica --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["nilotica"]
omit = ["tests/*", "hatch_build_tts.py", "hatch_build.py", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
